+++ packages/storage/src/implementation/type.ts
@@ -15,6 +15,8 @@
  * limitations under the License.
  */
 
+import { Code, FirebaseStorageError } from './error';
+
 /**
  * @return False if the object is undefined or null, true otherwise.
  */
@@ -31,16 +33,8 @@ export function isFunction(p: unknown): p is Function {
   return typeof p === 'function';
 }
 
-export function isObject(p: unknown): p is { [key: string]: unknown } | null {
-  return typeof p === 'object';
-}
-
-export function isNonNullObject(p: unknown): p is object {
-  return isObject(p) && p !== null;
-}
-
 export function isNonArrayObject(p: unknown): boolean {
-  return isObject(p) && !Array.isArray(p);
+  return typeof p === 'object' && !Array.isArray(p);
 }
 
 export function isString(p: unknown): p is string {
@@ -53,4 +47,24 @@ export function isNativeBlob(p: unknown): p is Blob {
 
 export function isNativeBlobDefined(): boolean {
   return typeof Blob !== 'undefined';
+}
+
+export function validateNumber(
+  argument: string,
+  minValue: number,
+  maxValue: number,
+  value: number
+): void {
+  if (value < minValue) {
+    throw new FirebaseStorageError(
+      Code.INVALID_ARGUMENT,
+      `Invalid value for '${argument}'. Expected ${minValue} or greater.`
+    );
+  }
+  if (value > maxValue) {
+    throw new FirebaseStorageError(
+      Code.INVALID_ARGUMENT,
+      `Invalid value for '${argument}'. Expected ${maxValue} or less.`
+    );
+  }
 }
